refactor(queue): extract data-retention pruning into its own module
Part of #4013's module-split sequence (step 7, after transient-locks.ts,
signal-snapshot.ts, duplicate-detection.ts, slop-detection.ts,
review-evasion.ts, and ci-resolution.ts): runRetentionPrune moves to
src/queue/retention.ts. Pure mechanical move, no behavior change -- a
re-export shim keeps this file's own internal caller and the existing
test/unit/retention.test.ts / test/unit/selfhost-pg-retention.test.ts
imports working unchanged.
